modified atoi, itoa, memccpy

+++ libft/ft_atoi.c
@@ -16,7 +16,7 @@ int	ft_atoi(const char *str)
 {
 	int i;
 	int sign;
-	int res;
+	long long res;
 
 	i = 0;
 	sign = 1;
